+++ doc/GSvar/changelog.md
@@ -1,5 +1,9 @@
 # GSvar change log
 
+###2018_06-34 (07.08.2018)
+- Filters: re-implemented filter concept
+- Report: add analyis pipeline version to report
+
 ###2018_06-18 (11.07.2018)
 - General: Added ancestry estimation.
 - Filters: The gene name filter now support using '*' as wildcard.
@@ -201,6 +205,7 @@
 --
 
 [back to main page](index.md)
+